fix: Set bond mode in host configuration
Used to be set in NewBondingHostConfig, which no longer exists in stprov
as of the transition to stboot's implementation of host configuration.

Note that bond name was not added by NewBondingHostConfig before, but it
is being written as of the refactoring by Kai.  This is the behavior we
want, but it stands out that stboot has worked without this until now.
A dependency of stboot likely defaults to "bond0", i.e., the same and
commonly used bond name that stprov sets (without being configurable).