test: Drop read/write host config efi tests
We used to have separate setup functions to produce and test valid host
configurations for different configurations: static, dhcp, and bonding.
The idea for them was: have something smaller that we can unit test.

Exactly how good these units tests were is a separate question.  Anyway.

Over time, these setup functions were not maintained.  Configuration
options were instead added directly in "main", and the logic to do a
correct setup are spread out all over the place.  As of the recent move
to using stboot's implementation of host configuration, these setup
functions were deemed obsolete and have been removed.  This commit also
removes the tests that were intended for these isolated setup functions.

Note that the internal reading and writing of EFI NVRAM is still being
tested separately from the removed unit test.  And that our qemu tests
try the path of json (un)marshalling with stboot's host configuration
package, then reading/writing with our internal EFI NVRAM reads/writes.
So in terms of actual test coverage we havn't really removed anything.

As an aside, the deleted tests did not run anymore.  The motivation for
deleting rather than fixing is (again) because they didn't add anything.

(What stprov really needs is refactoring to be more testable.  And this
is not the right time to get started with that type of refactoring.)