refactor: Parameterize QEMU's defaults (qemu.sh)
Still no changes that affect the actual testing, just preparing so that
we will be able to write better tests that do exactly what we think.

Note that the current test passes despite the gateway being wrong.  In
other words, stprov is guessing .1 but QEMU's default gateway is .2.  No
test fails because we are only communicating via a port forwarding rule.