fix: Add compile-time nob for setting template URL
Allow users to set their own template URL at compile time, which is a
way to type less (omitting -r URL in favor of -u USER and -p PASSWORD).

To be determined if this is the right way to have templated URLs.  This
is the old "mpt way", which we should think twice about before dropping.

Closes #43.  Related to #36.