test: Incorporate use of compiled-in defaults
This is somewhat important, because our Makefile will not fail if it is
trying to set a default Go variable that doesn't exist.  In other words,
if we change the name of a default variable without also updating the
Makefile correctly, it could come back to bite us and users later on.

Now we should catch this type of errors for all but the bond default.
It is considered out of scope because bonding is only tested manually.

Closes #18.