docs: Add tooling manual
The idea is to not write the things we have in our usage message twice
with slightly different wordings; and instead copy-paste from our
usage message to the manual which can then add a bit more detail.  The
resulting "man page" should be suitable for rendering on docs.sto.

By no means perfect, but likely modest overhead to maintain this manual
for a while without it getting out-of-sync.  It would likely be a good
idea to also specify "ST vX.Y.Z" or similar in the VERSION section,
which would be added by the script that pulls this into docs.sto.  The
pull script should also replace https links with docs.sto local files.
This would ensure we have consistent referencing within an ST release.

This commit drops the "cheat sheet", because it doesn't have to be
printed by stprov (add a print script to the initramfs instead).

This commit also makes a few minor edits to the usage message, to make
it fit a bit better when copy-pasting into relevant parts of the manual.