build: Allow slightly larger subnet by default
Nothing bad would happen even if we were to run with a much larger
subnet.  The worst case scenario is that our test machine takes entropy
from a random connection, and it requires us to actively click ENTER.

This makes it less likely we need to type the -a option on our SM.

Someone that knows the Glasklar topology may tighten this up later.